Bump version 1.5 codename "Canelo".
This release of VideoMorph has the purpose of honoring the memory of our dear friend Maikel Llamaret Heredia (a.k.a Canelo). He was one of the contributors to the project. He was also one of the more active contributors to Open Source and Free Software in Cuba. He was a dear friend and an exceptional human been. Canelo, a big hug to you from all the team of VideoMorph, we know you'll be okay. RIP Canelo.